Move helper functions to utility module and add config_override doco
