+++ src/utility.py
@@ -18,7 +18,8 @@ from pyspark.sql.types import StructType
 from constants import (
     SupportedSpecFormat,
     PipelineBundleSuffixesJson,
-    PipelineBundleSuffixesYaml
+    PipelineBundleSuffixesYaml,
+    FrameworkPaths,
 )
 
 
@@ -526,3 +527,42 @@ def set_logger(logger_name: str, log_level: str = "INFO") -> logging.Logger:
     logger.addHandler(console_output_handler)
 
     return logger
+
+
+def _has_visible_children(directory: str) -> bool:
+    """
+    Return True if `directory` exists and contains at least one child name not prefixed with `.`
+    """
+    if not os.path.isdir(directory):
+        return False
+    try:
+        names = os.listdir(directory)
+    except OSError:
+        return False
+    return any(not n.startswith(".") for n in names)
+
+
+def resolve_framework_config_path(framework_path: str) -> str:
+    """
+    Return FrameworkPaths.CONFIG_OVERRIDE_PATH when the config override path has at least one
+    non-hidden child and mirrors required ``./config`` layout; otherwise FrameworkPaths.CONFIG_PATH.
+
+    Raises:
+        FileNotFoundError: If override is non-empty but required paths are missing under config_override.
+    """
+    override_dir = os.path.join(framework_path, FrameworkPaths.CONFIG_OVERRIDE_PATH)
+    if not _has_visible_children(override_dir):
+        return FrameworkPaths.CONFIG_PATH
+
+    mapping_dir = os.path.join(override_dir, FrameworkPaths.DATAFLOW_SPEC_MAPPING)
+    global_paths = [
+        os.path.join(override_dir, name) for name in FrameworkPaths.GLOBAL_CONFIG
+    ]
+    if not os.path.isdir(mapping_dir) or not any(os.path.isfile(p) for p in global_paths):
+        raise FileNotFoundError(
+            f"Using {FrameworkPaths.CONFIG_OVERRIDE_PATH} requires both a global config file "
+            f"({' or '.join(FrameworkPaths.GLOBAL_CONFIG)}) and the "
+            f"{FrameworkPaths.DATAFLOW_SPEC_MAPPING} directory under ./config_override. "
+            f"Copy the full ./config tree into {FrameworkPaths.CONFIG_OVERRIDE_PATH}."
+        )
+    return FrameworkPaths.CONFIG_OVERRIDE_PATH
